improved nxUi
one improvement pending, before the HtmlViewerComponent was rendered using treeToHtml.
However, the component should be rendered by calling the individual toHtmls of the nodes of the component.
For that, the HtmlViewerComponent must be aware of the logic, which nodes are required for Html rendering and which are not. This is currently distributed in the HtmlRenders and the treeToHtml method. It might be required to create an extra class ResolvePolicy ... to bring this logic together in one place and make it accessible for the HtmlViewerComponent.