changing the way onImpossible events are reported from server to client.
this is to enable the client to react sensibly to some conditions, for instance when the owner of a node to be appended is not defined in the network, which might be due to that a previous message from the client has not arrived at the server yet. in this case, the client might try to resend the message.