Created a new test case in which multiple clients access one settings sub network.
The test case does not pass. This is caused by a fundamental bug in the AutoLoadNetwork. This network is supposed to synchronize nodes before they are accessed. However, since the operations on the network are synchronous but the synchronization operation is asynchronous, it cannot be guaranteed that the nodes are loaded before the operation is applied.

This can be mitigated in an JRE environment, where the thread accessing the network can be suspended, however, there exist no apparent workaround for the GWT environment.

Possible solutions are:
- cast an exception whenever a node is accessed, which has not been synchronized.
- provide a JRE auto loading network
- provide a special wrapping method to access networks. eg
Nx.assureLoaded(ForNetwork.check().ifNode(test).isDefined()).inNetwork(n).and(new WhenCompleted() { 
public void onCompleted(Boolean result) {
...
}
- reimplement Settings, Paths, Queries, ADTs, ... using a 'safer' approach accounting for remote networks.